+++ packages/mui-components/src/theme/components/MuiInputBase/index.ts
@@ -1,11 +1,20 @@
-import type { Components, Theme } from '@mui/material';
+import { alpha, type Components, type Theme } from '@mui/material';
 
 export const MuiInputBase: Components<Theme>['MuiInputBase'] = {
   styleOverrides: {
     root: ({ theme }) => ({
-      'border': `1px solid ${theme.palette.grey[800]}`,
+      'border': `1px solid ${theme.palette.divider}`,
       'padding': '0rem 1rem',
       'borderRadius': '4px',
+      'transition': '0.3s ease-out',
+
+      '&:hover': {
+        background: 'rgba(255, 255, 255, 0.1)',
+      },
+
+      '&:focus-within': {
+        background: 'rgba(255, 255, 255, 0.1)',
+      },
 
       '& .MuiOutlinedInput-notchedOutline': {
         border: 'none',