+++ packages/oidc/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @batoanng/oidc
 
+## 0.0.27
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.30
+
+## 0.0.26
+
+### Patch Changes
+
+- Updated dependencies
+  - @batoanng/mui-components@3.0.29
+
 ## 0.0.25
 
 ### Patch Changes